fix(fill): reference_spec fixture is required by fill, not by ver checker
This fixes up the previous commit that no longer includes the `spec_version_checker` plugin when executing `fill` or `execute` by removing it from their pytest-*ini files.
